move util -> deployment-util.js
